# c-deps: localize RocksDB's SSE4.2 usage to CRC32C