@tamird 2017-07-10
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/>

Regarding validating performance claims - the short answer is yes, but it's simpler and easier to just run relevant benchmarks yourself.

---

Reviewed 2 of 2 files at r1.
Review status: all files reviewed at latest revision, all discussions resolved, some commit checks failed.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/16949#-:-Kog8ie9AY35hi5aUR4a:b-9w3qb5)*
<!-- Sent from Reviewable.io -->