@tamird 2017-07-10
I think you might want to bump rocksdb-rebuild, too.