@benesch 2017-07-10
Yup, definitely going to upstream! The build system has changed enough between 5.5.1 and master, though, that it'll be two separate patches.

Any suggestions on which benchmarks to run?