@benesch 2017-07-11
Benchmarks from the first revision showed a noticeable slowdown. `objdump` revealed that GCC refuses to inline `FastCRC32` into its caller unless its caller is also marked with `__attribute__((feature("sse4.2"))`. Fixed by marking `ExtendImpl<FastCRC32>` with said attribute while leaving `ExtendImpl<SlowCRC32>` alone. 

A selection of benchmarks from `storage/engine`.

* HEAD~ vs this PR. Essentially no change; one op got slower, two got faster.

  ```
  $ benchstat bench.old bench.new
  name                                                      old time/op    new time/op    delta
  MVCCScan_RocksDB/rows=1/versions=10/valueSize=8-16          72.9µs ± 4%    73.0µs ± 3%    ~     (p=0.690 n=5+5)
  MVCCScan_RocksDB/rows=1/versions=10/valueSize=64-16         27.0µs ± 2%    27.3µs ± 3%    ~     (p=0.310 n=5+5)
  MVCCScan_RocksDB/rows=1/versions=10/valueSize=512-16        49.9µs ± 3%    50.5µs ± 4%    ~     (p=0.690 n=5+5)
  MVCCScan_RocksDB/rows=1/versions=100/valueSize=8-16          280µs ± 1%     280µs ± 2%    ~     (p=1.000 n=5+5)
  MVCCScan_RocksDB/rows=1/versions=100/valueSize=64-16        55.0µs ± 5%    55.6µs ± 3%    ~     (p=1.000 n=5+5)
  MVCCScan_RocksDB/rows=1/versions=100/valueSize=512-16       82.8µs ± 1%    83.5µs ± 3%    ~     (p=0.310 n=5+5)
  MVCCScan_RocksDB/rows=10/versions=10/valueSize=8-16          120µs ±10%     116µs ± 2%    ~     (p=0.421 n=5+5)
  MVCCScan_RocksDB/rows=10/versions=10/valueSize=64-16        73.4µs ± 6%    75.5µs ± 5%    ~     (p=0.095 n=5+5)
  MVCCScan_RocksDB/rows=10/versions=10/valueSize=512-16        124µs ± 5%     128µs ± 5%    ~     (p=0.095 n=5+5)
  MVCCScan_RocksDB/rows=10/versions=100/valueSize=8-16         395µs ± 2%     389µs ± 1%    ~     (p=0.421 n=5+5)
  MVCCScan_RocksDB/rows=10/versions=100/valueSize=64-16        144µs ± 4%     150µs ± 5%    ~     (p=0.310 n=5+5)
  MVCCScan_RocksDB/rows=10/versions=100/valueSize=512-16       280µs ± 3%     282µs ± 2%    ~     (p=0.548 n=5+5)
  MVCCScan_RocksDB/rows=100/versions=10/valueSize=8-16         533µs ± 3%     532µs ± 2%    ~     (p=1.000 n=5+5)
  MVCCScan_RocksDB/rows=100/versions=10/valueSize=64-16        511µs ± 5%     520µs ± 3%    ~     (p=0.310 n=5+5)
  MVCCScan_RocksDB/rows=100/versions=10/valueSize=512-16       874µs ± 4%     884µs ± 4%    ~     (p=0.690 n=5+5)
  MVCCScan_RocksDB/rows=100/versions=100/valueSize=8-16       1.45ms ± 2%    1.45ms ± 3%    ~     (p=1.000 n=5+5)
  MVCCScan_RocksDB/rows=100/versions=100/valueSize=64-16       982µs ± 2%    1038µs ± 4%  +5.74%  (p=0.008 n=5+5)
  MVCCScan_RocksDB/rows=100/versions=100/valueSize=512-16     2.21ms ± 2%    2.22ms ± 2%    ~     (p=0.841 n=5+5)
  MVCCScan_RocksDB/rows=1000/versions=10/valueSize=8-16       4.53ms ± 3%    4.56ms ± 3%    ~     (p=0.548 n=5+5)
  MVCCScan_RocksDB/rows=1000/versions=10/valueSize=64-16      4.57ms ± 2%    4.56ms ± 4%    ~     (p=0.841 n=5+5)
  MVCCScan_RocksDB/rows=1000/versions=10/valueSize=512-16     7.93ms ± 3%    8.05ms ± 7%    ~     (p=1.000 n=5+5)
  MVCCScan_RocksDB/rows=1000/versions=100/valueSize=8-16      11.9ms ± 1%    11.9ms ± 4%    ~     (p=1.000 n=5+5)
  MVCCScan_RocksDB/rows=1000/versions=100/valueSize=64-16     8.89ms ± 7%    9.46ms ± 4%    ~     (p=0.095 n=5+5)
  MVCCScan_RocksDB/rows=1000/versions=100/valueSize=512-16    21.0ms ± 2%    21.1ms ± 2%    ~     (p=0.841 n=5+5)
  IterOnBatch_RocksDB/writes=10-16                            1.98µs ± 3%    1.92µs ± 1%    ~     (p=0.056 n=5+5)
  IterOnBatch_RocksDB/writes=100-16                           2.53µs ± 7%    2.48µs ± 3%    ~     (p=0.548 n=5+5)
  IterOnBatch_RocksDB/writes=1000-16                          3.22µs ± 2%    3.14µs ± 2%  -2.44%  (p=0.032 n=5+5)
  IterOnBatch_RocksDB/writes=10000-16                         4.12µs ± 5%    4.09µs ± 3%    ~     (p=1.000 n=5+5)
  IterOnReadOnly_RocksDB/writes=10-16                         1.19µs ± 2%    1.18µs ± 3%    ~     (p=0.690 n=5+5)
  IterOnReadOnly_RocksDB/writes=100-16                        1.52µs ± 5%    1.47µs ± 3%    ~     (p=0.341 n=5+5)
  IterOnReadOnly_RocksDB/writes=1000-16                       1.68µs ± 2%    1.64µs ± 3%    ~     (p=0.222 n=5+5)
  IterOnReadOnly_RocksDB/writes=10000-16                      1.93µs ± 3%    1.92µs ± 5%    ~     (p=1.000 n=5+5)
  IterOnEngine_RocksDB/writes=10-16                           2.22µs ± 3%    2.19µs ± 4%    ~     (p=0.548 n=5+5)
  IterOnEngine_RocksDB/writes=100-16                          2.36µs ± 1%    2.37µs ± 2%    ~     (p=0.548 n=5+5)
  IterOnEngine_RocksDB/writes=1000-16                         2.60µs ± 2%    2.63µs ± 3%    ~     (p=0.151 n=5+5)
  IterOnEngine_RocksDB/writes=10000-16                        2.85µs ± 2%    2.87µs ± 3%    ~     (p=0.841 n=5+5)
  MVCCPut_RocksDB/valueSize=10-16                             5.58µs ± 7%    5.55µs ± 1%    ~     (p=0.421 n=5+5)
  MVCCPut_RocksDB/valueSize=100-16                            5.63µs ± 2%    5.62µs ± 1%    ~     (p=0.841 n=5+5)
  MVCCPut_RocksDB/valueSize=1000-16                           7.60µs ± 1%    7.58µs ± 3%    ~     (p=1.000 n=5+5)
  MVCCPut_RocksDB/valueSize=10000-16                          18.0µs ± 4%    17.7µs ± 2%    ~     (p=0.310 n=5+5)
  MVCCPutDelete_RocksDB-16                                    12.0µs ± 2%    12.2µs ± 4%    ~     (p=0.548 n=5+5)
  ClearIterRange_RocksDB-16                                    286ms ± 1%     285ms ± 0%    ~     (p=0.556 n=5+4)

  name                                                      old speed      new speed      delta
  MVCCScan_RocksDB/rows=1/versions=10/valueSize=8-16         110kB/s ± 0%   110kB/s ± 0%    ~     (all equal)
  MVCCScan_RocksDB/rows=1/versions=10/valueSize=64-16       2.37MB/s ± 2%  2.34MB/s ± 3%    ~     (p=0.333 n=5+5)
  MVCCScan_RocksDB/rows=1/versions=10/valueSize=512-16      10.3MB/s ± 3%  10.1MB/s ± 4%    ~     (p=0.595 n=5+5)
  MVCCScan_RocksDB/rows=1/versions=100/valueSize=8-16       30.0kB/s ± 0%  30.0kB/s ± 0%    ~     (all equal)
  MVCCScan_RocksDB/rows=1/versions=100/valueSize=64-16      1.17MB/s ± 5%  1.15MB/s ± 4%    ~     (p=1.000 n=5+5)
  MVCCScan_RocksDB/rows=1/versions=100/valueSize=512-16     6.18MB/s ± 1%  6.13MB/s ± 3%    ~     (p=0.286 n=5+5)
  MVCCScan_RocksDB/rows=10/versions=10/valueSize=8-16        670kB/s ± 9%   688kB/s ± 2%    ~     (p=0.762 n=5+5)
  MVCCScan_RocksDB/rows=10/versions=10/valueSize=64-16      8.73MB/s ± 5%  8.48MB/s ± 5%    ~     (p=0.095 n=5+5)
  MVCCScan_RocksDB/rows=10/versions=10/valueSize=512-16     41.3MB/s ± 5%  40.0MB/s ± 5%    ~     (p=0.095 n=5+5)
  MVCCScan_RocksDB/rows=10/versions=100/valueSize=8-16       204kB/s ± 3%   210kB/s ± 0%    ~     (p=0.238 n=5+4)
  MVCCScan_RocksDB/rows=10/versions=100/valueSize=64-16     4.44MB/s ± 4%  4.28MB/s ± 5%    ~     (p=0.246 n=5+5)
  MVCCScan_RocksDB/rows=10/versions=100/valueSize=512-16    18.3MB/s ± 3%  18.1MB/s ± 2%    ~     (p=0.548 n=5+5)
  MVCCScan_RocksDB/rows=100/versions=10/valueSize=8-16      1.50MB/s ± 3%  1.50MB/s ± 2%    ~     (p=0.865 n=5+5)
  MVCCScan_RocksDB/rows=100/versions=10/valueSize=64-16     12.5MB/s ± 4%  12.3MB/s ± 3%    ~     (p=0.310 n=5+5)
  MVCCScan_RocksDB/rows=100/versions=10/valueSize=512-16    58.6MB/s ± 4%  58.0MB/s ± 5%    ~     (p=0.738 n=5+5)
  MVCCScan_RocksDB/rows=100/versions=100/valueSize=8-16      550kB/s ± 2%   550kB/s ± 4%    ~     (p=1.000 n=5+5)
  MVCCScan_RocksDB/rows=100/versions=100/valueSize=64-16    6.52MB/s ± 2%  6.17MB/s ± 4%  -5.31%  (p=0.008 n=5+5)
  MVCCScan_RocksDB/rows=100/versions=100/valueSize=512-16   23.1MB/s ± 2%  23.1MB/s ± 2%    ~     (p=0.841 n=5+5)
  MVCCScan_RocksDB/rows=1000/versions=10/valueSize=8-16     1.77MB/s ± 3%  1.76MB/s ± 3%    ~     (p=0.714 n=5+5)
  MVCCScan_RocksDB/rows=1000/versions=10/valueSize=64-16    14.0MB/s ± 2%  14.0MB/s ± 4%    ~     (p=0.841 n=5+5)
  MVCCScan_RocksDB/rows=1000/versions=10/valueSize=512-16   64.6MB/s ± 3%  63.7MB/s ± 6%    ~     (p=1.000 n=5+5)
  MVCCScan_RocksDB/rows=1000/versions=100/valueSize=8-16     674kB/s ± 2%   672kB/s ± 3%    ~     (p=1.000 n=5+5)
  MVCCScan_RocksDB/rows=1000/versions=100/valueSize=64-16   7.21MB/s ± 7%  6.77MB/s ± 4%    ~     (p=0.095 n=5+5)
  MVCCScan_RocksDB/rows=1000/versions=100/valueSize=512-16  24.4MB/s ± 2%  24.2MB/s ± 2%    ~     (p=0.730 n=5+5)
  MVCCPut_RocksDB/valueSize=10-16                           1.80MB/s ± 6%  1.80MB/s ± 1%    ~     (p=0.476 n=5+5)
  MVCCPut_RocksDB/valueSize=100-16                          17.8MB/s ± 2%  17.8MB/s ± 1%    ~     (p=0.841 n=5+5)
  MVCCPut_RocksDB/valueSize=1000-16                          132MB/s ± 1%   132MB/s ± 3%    ~     (p=1.000 n=5+5)
  MVCCPut_RocksDB/valueSize=10000-16                         556MB/s ± 4%   565MB/s ± 2%    ~     (p=0.310 n=5+5)
  ClearIterRange_RocksDB-16                                  235MB/s ± 1%   234MB/s ± 3%    ~     (p=1.000 n=5+5)
  ```

* This PR vs. this PR without `-sse3`. Compiling without SSE3 is only slightly slower.

  ```
  benchstat bench.new bench.nosse3
  name                                                      old time/op    new time/op    delta
  MVCCScan_RocksDB/rows=1/versions=10/valueSize=8-16          73.0µs ± 3%    73.4µs ± 2%    ~     (p=0.841 n=5+5)
  MVCCScan_RocksDB/rows=1/versions=10/valueSize=64-16         27.3µs ± 3%    27.2µs ± 2%    ~     (p=0.841 n=5+5)
  MVCCScan_RocksDB/rows=1/versions=10/valueSize=512-16        50.5µs ± 4%    50.4µs ± 3%    ~     (p=1.000 n=5+5)
  MVCCScan_RocksDB/rows=1/versions=100/valueSize=8-16          280µs ± 2%     282µs ± 1%    ~     (p=0.056 n=5+5)
  MVCCScan_RocksDB/rows=1/versions=100/valueSize=64-16        55.6µs ± 3%    55.7µs ± 5%    ~     (p=0.841 n=5+5)
  MVCCScan_RocksDB/rows=1/versions=100/valueSize=512-16       83.5µs ± 3%    82.3µs ± 2%    ~     (p=0.222 n=5+5)
  MVCCScan_RocksDB/rows=10/versions=10/valueSize=8-16          116µs ± 2%     116µs ± 2%    ~     (p=1.000 n=5+5)
  MVCCScan_RocksDB/rows=10/versions=10/valueSize=64-16        75.5µs ± 5%    73.3µs ± 1%  -3.02%  (p=0.032 n=5+5)
  MVCCScan_RocksDB/rows=10/versions=10/valueSize=512-16        128µs ± 5%     127µs ± 3%    ~     (p=0.548 n=5+5)
  MVCCScan_RocksDB/rows=10/versions=100/valueSize=8-16         389µs ± 1%     391µs ± 1%    ~     (p=0.841 n=5+5)
  MVCCScan_RocksDB/rows=10/versions=100/valueSize=64-16        150µs ± 5%     146µs ± 4%    ~     (p=0.421 n=5+5)
  MVCCScan_RocksDB/rows=10/versions=100/valueSize=512-16       282µs ± 2%     280µs ± 2%    ~     (p=0.690 n=5+5)
  MVCCScan_RocksDB/rows=100/versions=10/valueSize=8-16         532µs ± 2%     531µs ± 3%    ~     (p=1.000 n=5+5)
  MVCCScan_RocksDB/rows=100/versions=10/valueSize=64-16        520µs ± 3%     512µs ± 5%    ~     (p=0.690 n=5+5)
  MVCCScan_RocksDB/rows=100/versions=10/valueSize=512-16       884µs ± 4%     900µs ± 6%    ~     (p=0.690 n=5+5)
  MVCCScan_RocksDB/rows=100/versions=100/valueSize=8-16       1.45ms ± 3%    1.46ms ± 1%    ~     (p=0.690 n=5+5)
  MVCCScan_RocksDB/rows=100/versions=100/valueSize=64-16      1.04ms ± 4%    1.00ms ± 4%    ~     (p=0.095 n=5+5)
  MVCCScan_RocksDB/rows=100/versions=100/valueSize=512-16     2.22ms ± 2%    2.22ms ± 4%    ~     (p=1.000 n=5+5)
  MVCCScan_RocksDB/rows=1000/versions=10/valueSize=8-16       4.56ms ± 3%    4.56ms ± 5%    ~     (p=1.000 n=5+5)
  MVCCScan_RocksDB/rows=1000/versions=10/valueSize=64-16      4.56ms ± 4%    4.55ms ± 4%    ~     (p=1.000 n=5+5)
  MVCCScan_RocksDB/rows=1000/versions=10/valueSize=512-16     8.05ms ± 7%    7.99ms ± 6%    ~     (p=0.690 n=5+5)
  MVCCScan_RocksDB/rows=1000/versions=100/valueSize=8-16      11.9ms ± 4%    12.0ms ± 3%    ~     (p=0.690 n=5+5)
  MVCCScan_RocksDB/rows=1000/versions=100/valueSize=64-16     9.46ms ± 4%    9.04ms ± 7%    ~     (p=0.095 n=5+5)
  MVCCScan_RocksDB/rows=1000/versions=100/valueSize=512-16    21.1ms ± 2%    21.4ms ± 4%    ~     (p=0.421 n=5+5)
  IterOnBatch_RocksDB/writes=10-16                            1.92µs ± 1%    2.00µs ± 2%  +4.17%  (p=0.008 n=5+5)
  IterOnBatch_RocksDB/writes=100-16                           2.48µs ± 3%    2.47µs ± 2%    ~     (p=0.794 n=5+5)
  IterOnBatch_RocksDB/writes=1000-16                          3.14µs ± 2%    3.19µs ± 4%    ~     (p=0.310 n=5+5)
  IterOnBatch_RocksDB/writes=10000-16                         4.09µs ± 3%    4.24µs ± 2%  +3.55%  (p=0.032 n=5+5)
  IterOnReadOnly_RocksDB/writes=10-16                         1.18µs ± 3%    1.19µs ± 3%    ~     (p=0.452 n=5+5)
  IterOnReadOnly_RocksDB/writes=100-16                        1.47µs ± 3%    1.43µs ± 5%    ~     (p=0.095 n=5+5)
  IterOnReadOnly_RocksDB/writes=1000-16                       1.64µs ± 3%    1.72µs ± 2%  +4.54%  (p=0.008 n=5+5)
  IterOnReadOnly_RocksDB/writes=10000-16                      1.92µs ± 5%    1.97µs ± 3%    ~     (p=0.222 n=5+5)
  IterOnEngine_RocksDB/writes=10-16                           2.19µs ± 4%    2.24µs ± 3%    ~     (p=0.167 n=5+5)
  IterOnEngine_RocksDB/writes=100-16                          2.37µs ± 2%    2.44µs ± 1%  +2.73%  (p=0.016 n=5+5)
  IterOnEngine_RocksDB/writes=1000-16                         2.63µs ± 3%    2.63µs ± 3%    ~     (p=1.000 n=5+5)
  IterOnEngine_RocksDB/writes=10000-16                        2.87µs ± 3%    2.89µs ± 1%    ~     (p=0.421 n=5+5)
  MVCCPut_RocksDB/valueSize=10-16                             5.55µs ± 1%    5.57µs ± 1%    ~     (p=0.690 n=5+5)
  MVCCPut_RocksDB/valueSize=100-16                            5.62µs ± 1%    5.82µs ± 2%  +3.51%  (p=0.008 n=5+5)
  MVCCPut_RocksDB/valueSize=1000-16                           7.58µs ± 3%    7.89µs ± 2%  +4.03%  (p=0.032 n=5+5)
  MVCCPut_RocksDB/valueSize=10000-16                          17.7µs ± 2%    18.4µs ± 1%  +4.02%  (p=0.008 n=5+5)
  MVCCPutDelete_RocksDB-16                                    12.2µs ± 4%    12.5µs ± 1%    ~     (p=0.151 n=5+5)
  ClearIterRange_RocksDB-16                                    285ms ± 0%     285ms ± 2%    ~     (p=0.730 n=4+5)

  name                                                      old speed      new speed      delta
  MVCCScan_RocksDB/rows=1/versions=10/valueSize=8-16         110kB/s ± 0%   110kB/s ± 0%    ~     (all equal)
  MVCCScan_RocksDB/rows=1/versions=10/valueSize=64-16       2.34MB/s ± 3%  2.35MB/s ± 2%    ~     (p=0.984 n=5+5)
  MVCCScan_RocksDB/rows=1/versions=10/valueSize=512-16      10.1MB/s ± 4%  10.2MB/s ± 3%    ~     (p=1.000 n=5+5)
  MVCCScan_RocksDB/rows=1/versions=100/valueSize=8-16       30.0kB/s ± 0%  30.0kB/s ± 0%    ~     (all equal)
  MVCCScan_RocksDB/rows=1/versions=100/valueSize=64-16      1.15MB/s ± 4%  1.15MB/s ± 5%    ~     (p=0.810 n=5+5)
  MVCCScan_RocksDB/rows=1/versions=100/valueSize=512-16     6.13MB/s ± 3%  6.22MB/s ± 2%    ~     (p=0.206 n=5+5)
  MVCCScan_RocksDB/rows=10/versions=10/valueSize=8-16        688kB/s ± 2%   690kB/s ± 1%    ~     (p=1.000 n=5+5)
  MVCCScan_RocksDB/rows=10/versions=10/valueSize=64-16      8.48MB/s ± 5%  8.74MB/s ± 1%  +3.02%  (p=0.032 n=5+5)
  MVCCScan_RocksDB/rows=10/versions=10/valueSize=512-16     40.0MB/s ± 5%  40.5MB/s ± 3%    ~     (p=0.548 n=5+5)
  MVCCScan_RocksDB/rows=10/versions=100/valueSize=8-16       210kB/s ± 0%   204kB/s ± 3%    ~     (p=0.095 n=4+5)
  MVCCScan_RocksDB/rows=10/versions=100/valueSize=64-16     4.28MB/s ± 5%  4.40MB/s ± 4%    ~     (p=0.341 n=5+5)
  MVCCScan_RocksDB/rows=10/versions=100/valueSize=512-16    18.1MB/s ± 2%  18.3MB/s ± 2%    ~     (p=0.587 n=5+5)
  MVCCScan_RocksDB/rows=100/versions=10/valueSize=8-16      1.50MB/s ± 2%  1.51MB/s ± 3%    ~     (p=0.968 n=5+5)
  MVCCScan_RocksDB/rows=100/versions=10/valueSize=64-16     12.3MB/s ± 3%  12.5MB/s ± 6%    ~     (p=0.690 n=5+5)
  MVCCScan_RocksDB/rows=100/versions=10/valueSize=512-16    58.0MB/s ± 5%  57.0MB/s ± 7%    ~     (p=0.690 n=5+5)
  MVCCScan_RocksDB/rows=100/versions=100/valueSize=8-16      550kB/s ± 4%   550kB/s ± 0%    ~     (p=0.643 n=5+5)
  MVCCScan_RocksDB/rows=100/versions=100/valueSize=64-16    6.17MB/s ± 4%  6.39MB/s ± 4%    ~     (p=0.095 n=5+5)
  MVCCScan_RocksDB/rows=100/versions=100/valueSize=512-16   23.1MB/s ± 2%  23.0MB/s ± 4%    ~     (p=1.000 n=5+5)
  MVCCScan_RocksDB/rows=1000/versions=10/valueSize=8-16     1.76MB/s ± 3%  1.76MB/s ± 6%    ~     (p=1.000 n=5+5)
  MVCCScan_RocksDB/rows=1000/versions=10/valueSize=64-16    14.0MB/s ± 4%  14.1MB/s ± 4%    ~     (p=1.000 n=5+5)
  MVCCScan_RocksDB/rows=1000/versions=10/valueSize=512-16   63.7MB/s ± 6%  64.2MB/s ± 6%    ~     (p=0.690 n=5+5)
  MVCCScan_RocksDB/rows=1000/versions=100/valueSize=8-16     672kB/s ± 3%   666kB/s ± 2%    ~     (p=0.587 n=5+5)
  MVCCScan_RocksDB/rows=1000/versions=100/valueSize=64-16   6.77MB/s ± 4%  7.09MB/s ± 7%    ~     (p=0.095 n=5+5)
  MVCCScan_RocksDB/rows=1000/versions=100/valueSize=512-16  24.2MB/s ± 2%  23.9MB/s ± 4%    ~     (p=0.460 n=5+5)
  MVCCPut_RocksDB/valueSize=10-16                           1.80MB/s ± 1%  1.80MB/s ± 1%    ~     (p=0.635 n=5+5)
  MVCCPut_RocksDB/valueSize=100-16                          17.8MB/s ± 1%  17.2MB/s ± 2%  -3.39%  (p=0.008 n=5+5)
  MVCCPut_RocksDB/valueSize=1000-16                          132MB/s ± 3%   127MB/s ± 2%  -3.90%  (p=0.032 n=5+5)
  MVCCPut_RocksDB/valueSize=10000-16                         565MB/s ± 2%   543MB/s ± 1%  -3.89%  (p=0.008 n=5+5)
  ClearIterRange_RocksDB-16                                  234MB/s ± 3%   236MB/s ± 2%    ~     (p=0.421 n=5+5)
  ```

* As a sanity check, this PR vs. a patch that removes the global SSE3 *and* targeted SSE4.2 optimizations. Much slower, as expected.

  ```
  $ benchstat bench.new bench.noopt
  name                                                      old time/op    new time/op    delta
  MVCCScan_RocksDB/rows=1/versions=10/valueSize=8-16          73.0µs ± 3%    82.1µs ± 1%   +12.42%  (p=0.008 n=5+5)
  MVCCScan_RocksDB/rows=1/versions=10/valueSize=64-16         27.3µs ± 3%    53.6µs ± 2%   +96.08%  (p=0.008 n=5+5)
  MVCCScan_RocksDB/rows=1/versions=10/valueSize=512-16        50.5µs ± 4%    96.7µs ± 2%   +91.30%  (p=0.008 n=5+5)
  MVCCScan_RocksDB/rows=1/versions=100/valueSize=8-16          280µs ± 2%     328µs ± 2%   +17.32%  (p=0.008 n=5+5)
  MVCCScan_RocksDB/rows=1/versions=100/valueSize=64-16        55.6µs ± 3%   121.3µs ± 1%  +118.35%  (p=0.008 n=5+5)
  MVCCScan_RocksDB/rows=1/versions=100/valueSize=512-16       83.5µs ± 3%   180.1µs ± 2%  +115.72%  (p=0.008 n=5+5)
  MVCCScan_RocksDB/rows=10/versions=10/valueSize=8-16          116µs ± 2%     129µs ± 6%   +10.91%  (p=0.008 n=5+5)
  MVCCScan_RocksDB/rows=10/versions=10/valueSize=64-16        75.5µs ± 5%   102.7µs ± 2%   +36.03%  (p=0.008 n=5+5)
  MVCCScan_RocksDB/rows=10/versions=10/valueSize=512-16        128µs ± 5%     197µs ± 3%   +53.74%  (p=0.008 n=5+5)
  MVCCScan_RocksDB/rows=10/versions=100/valueSize=8-16         389µs ± 1%     443µs ± 1%   +13.77%  (p=0.008 n=5+5)
  MVCCScan_RocksDB/rows=10/versions=100/valueSize=64-16        150µs ± 5%     254µs ± 6%   +69.51%  (p=0.008 n=5+5)
  MVCCScan_RocksDB/rows=10/versions=100/valueSize=512-16       282µs ± 2%     597µs ± 2%  +111.60%  (p=0.008 n=5+5)
  MVCCScan_RocksDB/rows=100/versions=10/valueSize=8-16         532µs ± 2%     545µs ± 4%      ~     (p=0.310 n=5+5)
  MVCCScan_RocksDB/rows=100/versions=10/valueSize=64-16        520µs ± 3%     560µs ± 8%    +7.79%  (p=0.008 n=5+5)
  MVCCScan_RocksDB/rows=100/versions=10/valueSize=512-16       884µs ± 4%    1205µs ± 4%   +36.34%  (p=0.008 n=5+5)
  MVCCScan_RocksDB/rows=100/versions=100/valueSize=8-16       1.45ms ± 3%    1.59ms ± 1%    +9.42%  (p=0.008 n=5+5)
  MVCCScan_RocksDB/rows=100/versions=100/valueSize=64-16      1.04ms ± 4%    1.47ms ± 4%   +41.33%  (p=0.008 n=5+5)
  MVCCScan_RocksDB/rows=100/versions=100/valueSize=512-16     2.22ms ± 2%    4.76ms ± 3%  +114.59%  (p=0.008 n=5+5)
  MVCCScan_RocksDB/rows=1000/versions=10/valueSize=8-16       4.56ms ± 3%    4.67ms ± 3%      ~     (p=0.095 n=5+5)
  MVCCScan_RocksDB/rows=1000/versions=10/valueSize=64-16      4.56ms ± 4%    4.89ms ± 3%    +7.09%  (p=0.008 n=5+5)
…